+++ components/script/dom/validitystate.rs
@@ -10,6 +10,7 @@ use dom_struct::dom_struct;
 use itertools::Itertools;
 use style_traits::dom::ElementState;
 
+use super::bindings::codegen::Bindings::ElementInternalsBinding::ValidityStateFlags;
 use crate::dom::bindings::cell::{DomRefCell, Ref};
 use crate::dom::bindings::codegen::Bindings::ValidityStateBinding::ValidityStateMethods;
 use crate::dom::bindings::inheritance::Castable;
@@ -225,5 +226,42 @@ impl ValidityStateMethods for ValidityState {
     // https://html.spec.whatwg.org/multipage/#dom-validitystate-valid
     fn Valid(&self) -> bool {
         self.invalid_flags().is_empty()
+    }
+}
+
+impl From<&ValidityStateFlags> for ValidationFlags {
+    fn from(flags: &ValidityStateFlags) -> Self {
+        let mut bits = ValidationFlags::empty();
+        if flags.valueMissing {
+            bits |= ValidationFlags::VALUE_MISSING;
+        }
+        if flags.typeMismatch {
+            bits |= ValidationFlags::TYPE_MISMATCH;
+        }
+        if flags.patternMismatch {
+            bits |= ValidationFlags::PATTERN_MISMATCH;
+        }
+        if flags.tooLong {
+            bits |= ValidationFlags::TOO_LONG;
+        }
+        if flags.tooShort {
+            bits |= ValidationFlags::TOO_SHORT;
+        }
+        if flags.rangeUnderflow {
+            bits |= ValidationFlags::RANGE_UNDERFLOW;
+        }
+        if flags.rangeOverflow {
+            bits |= ValidationFlags::RANGE_OVERFLOW;
+        }
+        if flags.stepMismatch {
+            bits |= ValidationFlags::STEP_MISMATCH;
+        }
+        if flags.badInput {
+            bits |= ValidationFlags::BAD_INPUT;
+        }
+        if flags.customError {
+            bits |= ValidationFlags::CUSTOM_ERROR;
+        }
+        bits
     }
 }
